sib_tools/requirements.txt
requests
keyring
keyrings.cryptfile
beaupy
regex
python-dotenv
traitlets
google-auth
google-api-python-client
jsonpatch
tabulate
boto3
flask
beautifulsoup4
gunicorn
dkimpy
